@@ -9,25 +9,25 @@
     "server": "npm run build && NODE_ENV=production node server.js"
   },
   "dependencies": {
-    "framer-motion": "^4.1.2",
-    "gray-matter": "^4.0.2",
+    "@datocms/cma-client-node": "^3.3.15",
+    "framer-motion": "^11.11.17",
+    "gray-matter": "^4.0.3",
     "isomorphic-fetch": "^3.0.0",
-    "leaflet": "^1.7.1",
-    "leaflet-defaulticon-compatibility": "^0.1.1",
-    "next": "canary",
-    "next-auth": "^3.13.3",
+    "leaflet": "^1.9.4",
+    "leaflet-defaulticon-compatibility": "^0.1.2",
+    "next": "^15.0.3",
+    "next-auth": "^4.24.10",
     "raw-loader": "^4.0.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-intersection-observer": "^8.31.0",
-    "react-leaflet": "^3.1.0",
-    "react-markdown": "^5.0.3",
-    "remark-unwrap-images": "^2.0.0",
+    "react-intersection-observer": "^9.13.1",
+    "react-leaflet": "^4.2.1",
+    "react-markdown": "^9.0.1",
+    "rehype-unwrap-images": "^1.0.0",
     "styled-components": "latest",
     "typeface-dosis": "^1.1.13"
   },
   "devDependencies": {
-    "babel-plugin-styled-components": "latest",
-    "globby": "^11.0.3"
+    "globby": "^14.0.2",
+    "react": "^18.3.1",
+    "react-dom": "18.2.0"
   }
 }
